Move FixedValueListenable into listenable.dart
