+++ packages/devtools_app/lib/src/screen.dart
@@ -7,8 +7,8 @@ import 'dart:async';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-import 'common_widgets.dart';
 import 'globals.dart';
+import 'listenable.dart';
 import 'scaffold.dart';
 import 'theme.dart';
